chore: add some tests for the table diff algorithm
