+++ apps/framework-cli/src/framework/core/mod.rs
@@ -25,6 +25,7 @@ pub mod execute;
 pub mod infra_reality_checker;
 pub mod infrastructure;
 pub mod infrastructure_map;
+pub mod partial_infrastructure_map;
 pub mod plan;
 pub mod plan_validator;
 pub mod primitive_map;
